Updated the model and added ref to pag_agreements. Refs #16577

+++ db/migrate/20250620231215_remove_acknowledged_users_id_from_acknowledgments.rb
@@ -1,7 +1,19 @@
 class RemoveAcknowledgedUsersIdFromAcknowledgments < ActiveRecord::Migration[7.1]
   def change
-    remove_foreign_key :acknowledgments, column: :acknowledged_users_id rescue nil
-    remove_index :acknowledgments, :acknowledged_users_id
-    remove_column :acknowledgments, :acknowledged_users_id
+    return unless table_exists?(:acknowledgments)
+     # Safely remove the foreign key if it exists
+    if foreign_key_exists?(:acknowledgments, column: :acknowledged_users_id)
+      remove_foreign_key :acknowledgments, column: :acknowledged_users_id
+    end
+
+    # Safely remove the index if it exists
+    if index_exists?(:acknowledgments, :acknowledged_users_id)
+      remove_index :acknowledgments, :acknowledged_users_id
+    end
+
+    # Safely remove the column if it exists
+    if column_exists?(:acknowledgments, :acknowledged_users_id)
+      remove_column :acknowledgments, :acknowledged_users_id
+    end
   end
 end
